fix(protocol): build the registration tests on Linux only
The tests referenced applicationsDir, desktopExecArg and desktopFileName,
which protocol_linux.go alone defines, so compiling the package's tests on
macOS or Windows failed with undefined symbols. Skipping on runtime.GOOS does
not help: the file still has to compile.

The Linux cases move to protocol_linux_test.go behind a build tag, which also
makes the runtime skips redundant. TestSchemeMatches stays platform-agnostic
because the invariant it checks holds everywhere.

`go build` ignores _test.go files, which is why a GOOS=darwin build looked
clean locally. `go vet` typechecks them and reproduces it.
